Fix issue #3
RISC.progOrg should be a dynamic quantity, not a constant, since it
depends on the size of the VAR area. A new exported variable
RISC.ProgOrg can be set from OSG, and then used in RISC.Load() to
relocate the code correctly.

Note: ProgOrg is still in bytes

Memory size increased from 128 words to 256 words.